Basic compile-time validation maybe working?
This is very hacky, and took a lot of iteration.

The basic idea is that, for each type, we generate a 'shadow type';
for each field on the base type, the shadow type has an
associated function with the same name, which returns the generate
shadow type for that field. Finally the shadow types have a method
that allows them to create keypaths to their concrete type.

Stringing this together: at compile time, we walk the keypath through
the shadow types; when we reach the end we know we have a valid
keypath, and we then concretize it.